Move state accounts into separate files
